ACPI / scan: Do not treat the OBDA8723 device as serdev for now
Commit e361d1f85855 ("ACPI / scan: Fix enumeration for special UART
devices") makes the acpi subsys enumerate all devices with a SerialBus
resource as serdev-s rather then as platform-device.

This causes the /dev/tty# node for uarts referenced by a UartSerialBusV2
to disappear, which breaks support for bluetooth over uart which needs a
btattach from userspace and does not yet have an in kernel hci serdev
driver.

One example of a bluetooth over uart device which lacks such a driver and
thus no longer works is the RTL8723BS's bluetooth part, this commit adds
an exception for this device to acpi_is_serial_bus_slave(), so that it
gets treated as before, allowing it to work with the userspace solution
for now.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: youling257 <youling257@gmail.com>