staging: rtl8723bs: remove a couple of redundant initializations
The initialization of PwrCfgCmd is redundant as the value is never read
and it is being re-assigned to PwrSeqCmd[AryIdx] inside a loop, hence
it can be removed.  Also, the initialization of ie_ptr is redundant as
the value is never read and it is being re-assigned in either path of
an if statement, hence it can be removed too.

Cleans up clang warnings:
drivers/staging/rtl8723bs/hal/HalPwrSeqCmd.c:53:15: warning: Value
stored to 'PwrCfgCmd' during its initialization is never read
drivers/staging/rtl8723bs/os_dep/ioctl_linux.c:374:7: warning: Value
stored to 'ie_ptr' during its initialization is never read

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: GuneetAtwal <guneetatwal@gmail.com>